requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==45.0.7
dj-database-url==3.0.1
Django==5.2.6
django-allauth==65.11.2
django-cloudinary-storage==0.3.0
gunicorn==23.0.0
idna==3.10
joblib==1.5.2
mutagen==1.47.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.10.0
